refactor : interface 패키지 변경, 코드 formatting 수정
