+++ user/src/main/java/com/zerobase/user/controller/UserDto.java
@@ -29,8 +29,6 @@ public class UserDto {
                     .phoneNumber(phoneNumber)
                     .build();
         }
-
-
     }
-
+    
 }
