+++ user/src/main/java/com/zerobase/user/domain/repository/UserReader.java
@@ -1,10 +1,13 @@
-package com.zerobase.user.service;
+package com.zerobase.user.domain.repository;
 
 public interface UserReader {
 
     boolean isExistsLoginId(String loginId);
+
     boolean isExistsNickname(String nickname);
+
     boolean isExistsEmail(String email);
+
     boolean isExistsPhoneNumber(String phoneNumber);
