+++ user/src/main/java/com/zerobase/user/domain/repository/UserReaderImpl.java
@@ -1,6 +1,5 @@
 package com.zerobase.user.domain.repository;
 
-import com.zerobase.user.service.UserReader;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
